I don't know what the fuck is going on, it just comes out garbled no matter what I seem to do.

+++ src/internal.h
@@ -27,10 +27,10 @@ struct pitch_data{
   float *analyzed_magnitude;
   float *synthesized_frequency;
   float *synthesized_magnitude;
-  size_t framesize;
-  size_t oversampling;
-  size_t overlap;
-  size_t samplerate;
+  long framesize;
+  long oversampling;
+  long overlap;
+  long samplerate;
 };
 
 void free_pitch_data(struct pitch_data *data);
